requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
Pillow==10.3.0
streamlit==1.33.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
